fix: Move required env var validation to runtime from build-time (as well as some minor fixes)
