+++ .env.development
@@ -1,5 +1,5 @@
 NEXT_PUBLIC_AUTH0_DOMAIN=dreamkast.us.auth0.com
 NEXT_PUBLIC_AUTH0_CLIENT_ID=0cWWdpGt4CpWjHJ9QIHtPm5GrJLS25lz
+NEXT_PUBLIC_AUTH0_AUDIENCE=https://event.cloudnativedays.jp/
 NEXT_PUBLIC_BASE_PATH=/cnsec2022/ui
 NEXT_PUBLIC_API_BASE_URL=http://localhost:8080/
-NEXT_PUBLIC_AUDIENCE=https://event.cloudnativedays.jp/
